+++ modules/claims_api/spec/factories/auto_established_claims.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-require 'claims_api/special_issue_mapper'
+require 'claims_api/special_issue_mappers/bgs'
 
 FactoryBot.define do
   factory :auto_established_claim, class: 'ClaimsApi::AutoEstablishedClaim' do
@@ -16,10 +16,10 @@ FactoryBot.define do
     flashes { form_data.dig('veteran', 'flashes') }
     special_issues do
       if form_data['disabilities'].present? && form_data['disabilities'].first['specialIssues'].present?
-        mapper = ClaimsApi::SpecialIssueMapper.new
+        mapper = ClaimsApi::SpecialIssueMappers::Bgs.new
         [{ code: form_data['disabilities'].first['diagnosticCode'],
            name: form_data['disabilities'].first['name'],
-           special_issues: form_data['disabilities'].first['specialIssues'].map { |si| mapper.code_from_name(si) } }]
+           special_issues: form_data['disabilities'].first['specialIssues'].map { |si| mapper.code_from_name!(si) } }]
       else
         []
       end
